backend/requirements.txt
# Core Framework
Flask==3.0.0
gunicorn==21.2.0

# AI Orchestration
langgraph==0.2.16
langchain-google-genai==1.0.10
langchain-core>=0.2.27,<0.3

# Image Processing
Pillow==10.4.0

# Speech-to-Text
google-cloud-speech==2.26.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Utilities
python-dotenv==1.0.1
